backend/requirements.txt
# Core Django dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.1.0

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0

# Development tools
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Task queue (for later)
celery==5.3.4
redis==5.0.1

# Azure integration
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Utilities
python-decouple==3.8
requests==2.31.0